requirements/app.txt
galaxy.plugin.api==0.31
pyobjc-framework-CoreServices==5.1.2; sys_platform == 'darwin'
